+++ src/gql/interfaces/elements/Category.php
@@ -36,7 +36,7 @@ class Category extends Structure
     /**
      * @inheritdoc
      */
-    public static function getType($fields = null): Type
+    public static function getType(): Type
     {
         if ($type = GqlEntityRegistry::getEntity(self::getName())) {
             return $type;