+++ src/gql/interfaces/elements/Entry.php
@@ -38,7 +38,7 @@ class Entry extends Structure
     /**
      * @inheritdoc
      */
-    public static function getType($fields = null): Type
+    public static function getType(): Type
     {
         if ($type = GqlEntityRegistry::getEntity(self::getName())) {
             return $type;